+++ src/Ix.dom.asynciterable.ts
@@ -1,1 +1,2 @@
 export * from './asynciterable/index';
+import './add/asynciterable-operators/todomstream';
