+++ src/Ix.dom.iterable.ts
@@ -1,1 +1,2 @@
 export * from './iterable/index';
+import './add/iterable-operators/todomstream';
